Global variables finalization
- Moved responsibility of loading files specified in parameters to global object
-- controller.py load parameters then calls globals.init
-- globals.init handles all the loading (execept agents attributes and behaviors)
-- agent loading still inside simulation (todo: move this responsibilities to plugin loader

- controller.py now more focused on the control of the application as is custom in the MVC pattern
-- controler still loads the view (but is not a huge problem for now
